+++ use_class.py
@@ -1,9 +1,10 @@
-import requests, getpass
+import requests, getpass, urllib, json
 import vmware_horizon
 
 requests.packages.urllib3.disable_warnings()
 #url = input("URL\n")
-url="https://loftcbr01.loft.lab"
+#url="https://loftcbr01.loft.lab"
+url="https://pod2cbr1.loft.lab"
 #username = input("Username\n")
 username = "m_wouter"
 #domain = input("Domain\n")
@@ -13,9 +14,31 @@ pw = getpass.getpass()
 
 hvconnectionobj = vmware_horizon.Connection(username = username,domain = domain,password = pw,url = url)
 hvconnectionobj.hv_connect()
-#print(hvconnectionobj)
+# #print(hvconnectionobj)
+# obj = vmware_horizon.Inventory(url=hvconnectionobj.url, access_token=hvconnectionobj.access_token)
 obj = vmware_horizon.Inventory(url=hvconnectionobj.url, access_token=hvconnectionobj.access_token)
-sessions = obj.get_sessions(maxpagesize=1)
+# sessions = obj.get_sessions(maxpagesize=1)
+#print(obj.rds_servers())
+filter = {}
+filter["type"] = "And"
+filter["filters"] = []
+filter1={}
+filter1["type"] = "Contains"
+filter1["name"] = "name"
+filter1["value"] = "Pod02"
+print(filter1)
+filter["filters"].append(filter1)
+print(filter)
+tmp = json.dumps(filter,separators=(', ', ':'))
+print(tmp)
+filter_url = urllib.parse.quote(tmp)
+print(filter_url)
+# bla = urllib.parse.urlencode(filter)
+# print(bla)
+machines = obj.get_machines(maxpagesize=1, filter = filter)
+for i in machines:
+    print(i)
+
 
 
 # print(dc)